src/main/java/lv/javaguru/java1/student_tatjana_grigorcaka/lesson_11/homework/day_2/DayOfTheWeekDetectorSwitchVersion.java
package lv.javaguru.java1.student_tatjana_grigorcaka.lesson_11.homework.day_2;

public class DayOfTheWeekDetectorSwitchVersion implements DayOfTheWeekDetector {

    @Override
    public String detectDayName(int number) {
        String result = "";
        switch (number) {
            case 1:
                result = "Monday";
                break;
            case 2:
                result = "Tuesday";
                break;
            case 3:
                result = "Wednesday";
                break;
            case 4:
                result = "Thursday";
                break;
            case 5:
                result = "Friday";
                break;
            case 6:
                result = "Saturday";
                break;
            case 7:
                result = "Sunday";
                break;
            default:
                result = "Please input a valid number between 1 and 7!";
                break;
        }
        return result;
    }
}




